+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Properly parse the `<main>` tag.
 * Remove support for the non-standard `<command>` tag.
 * Create `HTMLCanvasElement` when parsing `<canvas>` tags.
+* Create `HTMLDialogElement` when parsing `<dialog>` tags.
 
 # domino 1.0.23 (30 Jan 2016)
 * Fix `CSSStyleDeclaration#setProperty`. (#71)